карта и активация/деактивация формы
