+++ js/main.js
@@ -1,14 +1,11 @@
-import {createSimilarOffers} from './data.js';
-// import {renderSimilarOffers} from './similar-elements.js';
-import {onSelectType, onSelectTime, inactivateForm, inactivateFields, inactivateMapForm, inactivateMapFormFields} from './form.js';
+// import {createSimilarOffers} from './data.js';
+// import {renderSimilarOffers} from './popup.js';
+import {onSelectType, onSelectTime, inactivateForm} from './form.js';
 import {initMap} from './map.js'
 
-const offers = createSimilarOffers(1);
+// const offers = createSimilarOffers(10);
 // renderSimilarOffers(offers);
 onSelectType();
 onSelectTime();
-inactivateFields();
 inactivateForm()
-inactivateMapForm();
-inactivateMapFormFields();
-initMap(offers);
+initMap();
